Added linear layout with two textareas to make lap pace more visible (uses long to show units)
Color will change if lap pace changes. NEed to do more testing with this to make colours change at right time

App was assuming that we use miles in GB. We do for driving, but most races are done in Kilometers. So I've made it use KM by default for Brits.

Todo: Make miles default for US. Low priority as I'm the only one using this and I'm not American. And it doesn't make sense that they are the only ones using outdated measurments.